fix(worker): normalize player region to short codes + TS-parity persistence
- provider.rs / profile_enrichment.rs: normalize_region maps long-form names
  (North America, Europe, Brazil, ...) to short codes the activity page
  REGION_COLORS recognizes; Latin America flags to Unknown, matching TS.
- discovery_store.rs: emit ASIA (not "Asia") so it doesn't render Unknown.
- match_facts.rs: upsert hourly_match_counts after ranked match persistence
  (TS upsertHourlyCount parity, queue 486 only, non-limited); persist bans
  for all populations, not just ranked (TS buffer-processor parity).
- maintenance.rs: rebuild match_compositions_ranked in derived-projections
  rebuild from source facts (TS derived-projection-tracker parity).
- nonranked_acquisition.rs: cast direct/roster player counts to i16 for
  int2 columns (fixes tokio-postgres serialization).
- requested_match.rs: derive Serialize/Deserialize on RequestedMatchStatus
  with snake_case rename for API payloads.

+++ src/backend-rust/src/workers/nonranked_acquisition.rs
@@ -185,7 +185,7 @@ impl NonrankedAcquisitionRepository {
             "UPDATE nonranked_match_acquisition SET status=$2,quality=$3,direct_player_count=$4,\
              roster_player_count=$5,roster_attempts=CASE WHEN $6 THEN roster_attempts+1 ELSE roster_attempts END,\
              terminal_reason=$7,error_message=NULL,lease_until=NULL,completed_at=now(),updated_at=now() WHERE match_id=$1",
-            &[&result.match_id, &result.state.database(), &quality, &direct_count, &roster_count,
+            &[&result.match_id, &result.state.database(), &quality, &(direct_count as i16), &(roster_count as i16),
               &attempted_roster, &result.terminal_reason],
         ).await?;
         transaction.commit().await?;
